state: added add/get deep equivalence test
The equivalence tested isn't necessarily correct, and
comparing private details is discouraged in the project.
The implementation might choose to cache information, or
to have different logic when adding or removing, and the
comparison might fail despite it being correct.
That said, we've had bugs with txn-revno being incorrect
before, so this testing at least ensures we're conscious
about such changes.