Add new preferred charm/bundle url format
Fixes http://pad.lv/1584193.

The new format is user/name/series/revision, with all of the fields but
name being optional.

In the code this is called the V3 format. V1 is the
cs:~user/series/name-revision format, and V2 are full jujucharms urls
such as https://jujucharms.com/u/aisrael/mysql-benchmark/trusty/5. The
code still supports the old formats for backwards compatibility.

To distinguish between the V2 url series/name and the V3 url user/name
we need to use a list of series from juju/itil/series.