LP 1640223: Add funcition that returns `actionNameRule`
Variable `actionNameRule` is needed inside juju's repository
packages. By directly importing the variable in the juju package,
the code losses readability due to the fact that it doesn't
specify the origin of the variable.
With a function the intent of importing the variable from a different
scope is more clear.