Add helper for splitting bundle into base and overlay-only parts
The helper relies on the fact that all BundleData (and friends) define
their fields with the omitempty modifier for all supported serialization
outputs (bson, json and yaml). This allows us to produce the two bundles
by duplicating the original bundle and zeroing the fields that we don't
need.

This way, when we serialize the resulting bundles we get two subsets
that when merged together yield the original bundle.