Use absolute path when creating error messages for missing bundle includes
When resolving missing includes for a base bundle or an overlay, the
generated error messages used to specify the relative include path to
the missing resource.

The emitted error was of no use to the end-user as it omitted a critical
piece of information: the *base* path for the bundle/overlay that was
used to resolve the relative include path.

This commit attempts to address this issue by modifying the
aforementioned error messages to include the absolute path to the
missing resource.