Charm origin
With the addition of different stores from which you can install a charm
from, we need to correctly model where a charm originates from. The
changes here allow just the origin source for now, I can see in the
future that more fields will want to be added.

The possible sources that we need to model are charmhub and everything
else. With this change though we can give feedback to the operator about
where a charm originates from to help them understand why a charm might
not be updated without switching or updating the path.

Without correctly modelling this we have to inspect the URL of a charm
and that's just more tricky. If we know when we deploy a charm what the
source was, we should store that.

---

The code is really simple, it's an optional addition to a charm, so
libraries using this won't have to do anything.