Expose rewriting of charmhub URLs
In theory charmhub URLs should never be written into the database as the
version 2 of the charmhub URL. Yet, we should do our utmost to ensure
that is the case. By exposing the re-write capabilities of our charm URL
function, we can re-use that with in Juju directly and ensure that all
charm URLs are of the correct fashion.