Allow charmhub url parsing
The following updates the charm URL code to allow the parsing of
charmhub prefix schema. This is a breaking change and will require a v8
of the charm package!

Essentially all charm URLs will require a schema prefix in every case.
Juju or other consumers will have to lift the raw url into the right
schema so that you get the right url.