Correctly handle opaque url paths
The way that the go url parser parses what should really be potentially
invalid urls, we have to check that the Opaque part of the URL struct is
not empty before using that as the path.

The charmhub url is now just a prefix (ch), with a name suffix. Nothing
else is allowed in the URL, as we're hoping to use the charm origin to
correctly model when something has changed.