Include ch schema in String().
If we do not include the 'ch' schema in the String(), a *charm.URL is
not saved to the juju db with the ch schema.  It appears that mongo is
saving via the String().  This may create issues down the road when a
new charm schema is created.

User facing messages in juju have been cleaned up so that
charm.URL.String() is no longer directly printed.