Remove legacy services
Services have been time and memorial, unfortunately they were
deprecated in version 2.x (2016 timeframe). I think authors of charms
have had enough time to move over to applications.

This removes services, yet we have to keep the horrid hack of
normalizing the data when we unmarshal. It's wrong and shouldn't be in
this code, it should be a two step process and remove the magic that
causes us issues in the long run.