Allow architecture and series in identifier URLs
Identifier URLs (charmhub URLs) are required to now identify themselves
with ability to adorn themselves with architectures and series. This is
so the uniquness of a URL within Juju is more than just the name and
revision. It allows more fine grained constraints about what a charm URL
is.

It does have the consequence that any other repsonsibilities outside of
the constrainted URL domain is now lost. If you want to use the charm
URL for a differing architecture becomes unwise or even impossible.