Application short syntax
The following implements application short syntax as the following:

    applications:
        mysql: {}

There are valid reasons for doing this. Initially it was thought that ~
could be used to indicate the short syntax, but ~ already indicates
null. This is fine, but causes all sorts of hell when integrating with
overlays. ~ becomes a way to delete charms and that makes is confusing
and very hard to work out if we should use the short syntax or delete a
charm.

The best way forward is to just side step this completely and use the
empty object type. It's not as short, but it's very clear from the code
how to move forward with it.

Additionally this also adds the default number of units to 1, which also
handles the Scale and NumUnits convertion. Essentially Scale can just be
ignored once it's been normalised.

Adding the default number to 1 causes a lot of pain with subordinates.
We don't know they're a subordinate until much later. To prevent an
error comming from the verification process, we return a map of warnings
that have been emitted from the verification process and if the warning
marries with a subordinate we should be able to work out if we should
output an error from that.

I think maybe in a following commit we should hold internal state about
if a value was patched to become 1 and if so only check that before
using the error and we might be able to remove warnings completely.