Prevent Format V1 and V2 mixing together
Metadata V2 should not be allowed into charmhub without the appropriate
manifest.yaml. To prevent this becoming the norm where V1 and V2 are
intermixed and therefore supported becomes a problem in the future about
modelling metadata.

Hyrum's Law states:

    With a sufficient number of users of an API,
    it does not matter what you promise in the contract:
    all observable behaviors of your system
    will be depended on by somebody.

This is going to happen and we should nip this in the bud very quickly.
If metadata v2 should be allowed on charmhub it should be done in the
right way by following the spec, or not at all!