Encode channel into a URL
The following encodes a channel into a URL. A channel is comprised of
the following parts; track, risk and branch. Track in this scenario is
always required and the other parts are optional.

To keep backwards compatibility with version 1 and HTTP version of the
charm URL, I have elected to keep the series. Series and channel are
exclusive and the String() method defines that. I belive we should
implement a validate for the URL to ensure that all components are
correct at the time of manual construction, but during parsing it will
be impossible to have series and channel at the same time.

One last thing, I'm not sure about the split on channel. As this is
intended to be a URL, I would like to keep it as close to a URL in terms
of semantics. So the URL should be parsable via url.Parse from the std
package, hence why we're not using #.