Encode base into charm URL
The following encodes a base into a URL. A base is comprised of
the following parts; track, risk, and branch. Track in this scenario is
always required and the other parts are optional.

To keep backwards compatibility with version 1 and HTTP version of the
charm URL, I have elected to keep the series. Series and base are
exclusive and the String() method defines that. I believe we should
implement a validate for the URL to ensure that all components are
correct at the time of manual construction, but during parsing, it will
be impossible to have series and base at the same time.

One last thing, I'm not sure about the split on the base. As this is
intended to be a URL, I would like to keep it as close to a URL in terms
of semantics. So the URL should be parsable via url.Parse from the std
package, hence why we're not using #.