Allow for backwards compatibility.
This is required as we're reading a url from a database and because the
codec for the value in the database has been placed on the type (why the
hell was that done?), we have no option but to support it here, in the
library and not at the database layer.

This is terrible in so many ways, but so little time to actually fix it.