Move bases from metadata to manifest.
Implementing metadata v2 spec changes, where bases do not live in
metadata.yaml, but are created by charmcraft and uploaded with the
charm.  Creates some issues whether the metadata in the charm is v1 or
v2 without knowledge of a manifest.  To be resolved in a near future
commit.

Implement ReadManifest. Update ReadCharmDir and ReadCharmArchive to
attempt reading the manifest, dependent of whether the charm is using
metadata v2.