Split and move metadata validation in ReadMetadata.
In ReadMetadata, ensure that the data is unambiguous in terms of
is it v1 or v2.

ReadCharm, will determine if the charm is v1 or v2, and call metadata
Check with the result.  Check was previously done in ReadMetadata.

Combine marshalling and unmarshalling into 1 method each.  Users of the
package should run Check before marshalling.

Remove unused Platform type and MustParseTerm method.