Imports base logic from juju/systems to this package.
Will be consumed by this package and juju in the future.  There are a
few important changes.

The base string does not return an interpretation of the series from
the base, it is simplpy a string representation of the base with slash
deliminators.

The name is changed to lower case where possible.  Issues have been seen
in juju making requests to the Charmhub API.

Use os pkg version of operating systems instead of copying over.