Resolve out of sync presence map
The following attempts to keep the presence map in-sync with the actual
data after a normalisation. During a normalisation "scale" is swapped
out for "num_units". This is fine, except when it comes to overlays you
now don't know if scale or num_units was used for the overlays and
subsequent scale calls are now not used.

This is actually a problem with the library itself. There should be an
internal representation of a bundle that carries all this important
information, which should include schemas etc. At the very end of the
reading/parsing process we should hand back a clean version of that
charm/bundle. That way we can make the bundle grow organically without
the requirement of so many hacks.

The code is simple, detect and resolve a normalisation and just insert
the right value into the presence map before applying an overlay.