Remove the panic
When dealing with deletions of applications for overlays, the presence
map will have removed applications. In that case we need to not panic
when attempting to resolve the fields for an out of sync map.

It is much safer to always return a map back and then determine if the
field present is on the empty FieldPresenceMap. We will fail a bit later
on, but it means we don't have to force consumers of the forField check
for existence of a value.