Manifest: Allow manifest to be nil
In order to correctly understand if the manifest is actually empty, we
need to not have an empty manifest for missing manifests. When parsing
the format we can actually check for the presence of manifest easily.

This does mean that you have to check for manifest, but in Juju, we're
doing that in most places.