Charm+format: Series always dictates a format v1
If a charm contains a series, even with the presence of manifest.yaml
then that charm is a v1. This is a damn shame, because our modelling
becomes weaker inside of juju, but it seems like charmcraft doesn't
restrict uploads with series and manifests.

The fix is a bit more involved, as we can use a set.String for better
look up times. It also removes duplications with in the selection
reasons.