Removes metrics code
We no longer support metrics code, so we should remove the code. Moving
it to dqlite would require understanding and maintaining code that has
not full tested or integrated with romulus or omnibus is sometime. Fully
understanding if all parts are correct would be challenging.

If in the future we want to reuse parts, we can always look back in the
history of the repo to bring back the parts that are required (if any).